# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search keyword matched against username, name, or national ID",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role",
            "schema": {
              "example": "teacher",
              "type": "string",
              "enum": [
                "admin",
                "registrar",
                "teacher",
                "coordinator"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1+)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size (1-100)",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsersDto"
              },
              "examples": {
                "default": {
                  "summary": "Create user",
                  "value": {
                    "nationalId": "199001011234",
                    "username": "jdoe",
                    "passwordHash": "$2b$10$FDSf0rjLQ8HsZb0zFvYeOeZKz3R8G5UfH6OteIFqiyIqOQUd0pD3e",
                    "role": "teacher",
                    "firstName": "John",
                    "lastName": "Doe",
                    "email": "john.doe@example.edu",
                    "phone": "+1-202-555-0199",
                    "isActive": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsersDto"
              },
              "examples": {
                "default": {
                  "summary": "Update user contact information",
                  "value": {
                    "email": "johnny.doe@example.edu",
                    "phone": "+1-202-555-0100",
                    "isActive": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "School Man API",
    "description": "Generated OpenAPI specification for School Man backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUsersDto": {
        "type": "object",
        "properties": {
          "nationalId": {
            "type": "string",
            "example": "199001011234",
            "description": "Unique national identifier"
          },
          "username": {
            "type": "string",
            "example": "jdoe",
            "description": "Unique username for login"
          },
          "passwordHash": {
            "type": "string",
            "example": "$2b$10$FDSf0rjLQ8HsZb0zFvYeOeZKz3R8G5UfH6OteIFqiyIqOQUd0pD3e",
            "description": "BCrypt password hash"
          },
          "role": {
            "type": "string",
            "example": "teacher",
            "enum": [
              "admin",
              "registrar",
              "teacher",
              "coordinator"
            ],
            "description": "Role assigned to the user"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name of the user"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.edu",
            "description": "Contact email address"
          },
          "phone": {
            "type": "string",
            "example": "+1-202-555-0199",
            "description": "Contact phone number"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the user is active"
          }
        },
        "required": [
          "nationalId",
          "username",
          "passwordHash",
          "role"
        ]
      },
      "UpdateUsersDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}